+++ integration-tests/instrumentation-task-disabled/src/main/kotlin/MainKt.kt
@@ -1,0 +1,8 @@
+class MainKt {
+  companion object {
+    @JvmStatic
+    fun main(args: Array<String>) {
+      println("Hi!")
+    }
+  }
+}