+++ integration-tests/instrumentation-task/src/main/java/Main.java
@@ -1,0 +1,7 @@
+import org.jetbrains.annotations.NotNull;
+
+class Main {
+    public static void main(@NotNull String[] args) {
+        System.out.println("Hello, World!");
+    }
+}